requirements.txt
fastapi
uvicorn
jinja2
python-multipart
cryptography
pycryptodome
python-dotenv
